Update heavy.go
Typed this straight in the browser so I am not sure if it complies, but I think this is all you need. Axt also contains a query, but I think we if implement functions using the intervals of the target/reference. We can all the same functions for the query if we implement a swap target and query (Chr Start End)