Update README.md installation instructions
From Dan's comment "Just a heads up, the install directions for gonomics are out of date. go get no longer works. I think we should advise people to clone the repo and run go install ./…. I may work on updating it later in the week, but if others are motivated then have at it."